finish card atom and molecules
