+++ src/components/molecules/testimonial-card/component.tsx
@@ -1,7 +1,18 @@
+import Card from '../../atoms/card';
 import React from 'react';
 
-export const TestimonialCard: React.FC = () => (
-  <div>
-    {/* Placeholder for TestimonialCard component */}
-  </div>
+type Variant = 'primary' | 'secondary' | 'tertiary' | "ghost";
+
+interface Props {
+  text: string;
+  name: string;
+  job: string;
+}
+
+export const TestimonialCard: React.FC<Props> = ({ text, name, job }) => (
+    <Card className="bg-white rounded-xl shadow-xl p-6 pt-20 pb-20">
+      <h2 className="italic text-1xl font-small mb-2">{text}</h2>
+      <p className="text-gray-400 pt-5 font-bold">{name}</p>
+      <p className="italic text-gray-500">{job}</p>
+  </Card>
 );
